ref(samples): [Queue Instrumentation 18] Move Kafka sources into queues.kafka package
Move KafkaConsumer and KafkaController in the three Spring Boot Jakarta
samples (jakarta, jakarta-opentelemetry, jakarta-opentelemetry-noagent)
into a queues.kafka sub-package.

No behavior change. Groups the Kafka-specific sample sources so future
queue integrations can sit next to them under queues.
